Use GitHub Releases for official builds
This get ride of AppVeyor for official builds once and for
all.

As people have been using AppVeyor for custom build and as number might
clash with official releases, the minor version of the project was
incremented.

AppVeyor will still be used for PRs.

Also to avoid more confusion between AppVeyor builds and official one,
AppVeyor builds are now postfixed with "-prXXX".