input: Fixes TouchPoint wrong attribute
Even if Touch works in some games, we didn't set the `End` attribute and then the wrong value is written in the shared memory. That lead to some touch issues in some specific games.

It could be nice to get some feedbacks, if anyone knows games with touch issues feel free to test if that fixes it.

Fixes #2351
Now Mini Metro (and probably more games) is playable: