Vulkan: Device map buffers written more than flushed
This PR retunes the buffer migration system to map buffers that are written significantly more than they are read back on device, rather than on host.

NVIDIA GPUs seem to have a lot of trouble mapping a storage buffer for write in a ton of places throughout a frame, probably because it has to ensure coherency every time. Unfortunately we can't seem to disable that, so using host buffers always has a large hidden cost when using a storage buffer in a lot of places. TOTK does a lot of storage writes from _fragment_ of all things, in many different places.

This causes a small performance regression in BOTW, but a big performance boost in TOTK. This is due to reading from device memory being uncached. There may be a better solution to this problem.