Log file: append the current process ID to make name unique
If a user opens Ryujinx 2+ times quickly  in a row, the emulator will crash. This is because the log file has same name and it will cause the 2nd instance crash due to "being used by another process".
Since process ID is always unique, so this simple PR will resolve the race condition outlined in #5172.