Ryujinx.Ava: Fix menus on gamescope
gamescope seems to only support rendering one surface window at a time.

This works fine for our GTK frontend as everything is embedded in the
same window.

Avalonia on the other hand will attempt to create a new window for every
menu, making them not render properly.

Adding insult to the injury, gamescope seems to ignore the surface
update for those Avalonia's menus and refuse to present them.

To workaround all of this, we switch OverlayPopups to true to ensure we embed all popups in the
window.

Note that all our current dialogs (about us, warning about keys,...)
will still show on top without the main window behind because of
gamescope behavior.

Signed-off-by: Mary <mary@mary.zone>