Add raw screenshot gamepad button
This triggers the same raw screenshot as the keyboard hotkey,
i.e. the game is not informed of the intent to screenshot,
so it cannot add watermarks etc.

The button can be configured using both the regular UI and GTK.

Includes English and German translations.

An alternate approach is proposed in draft PR #4503:
Allow triggering all hotkeys using the gamepad.

Re: #5952